Reduce excerpt length. Last minute fix.
I just pushed the site up to find that to my dismay the excerpt will not close
pre tags that happen to start before the excerpt chop-off point. This means that
the entire home page was rendered as if it was in a pre tag. That's bad. So this
is a quick fix. I should come back and add logic to close pre tags so this
doens't happen again.