Remove show-hide event handlers
Turns out we can completely handle this sort of state using hover. The
event handlers are only necessary when we need an event from one el to
affect another el. In our case so far almost everything could be
handled with hover.

I do still have a selectAll event handler in place to select the text
box when the user focuses it, although this could arguably be done with
inline JS.