Normalize frontmatter across all posts and add created date
This is a sweeping change, but it is meant only to standardize some front matter
information in posts. Namely, published posts should all have a created and
published date represented in ISO form.

There is still a "date" key created (i think) by hexo which is now redundant,
but I haven't yet removed it.